refactor: move acess-token to authentication
