+++ src/apps/backend/modules/account/rest_api/account_view.py
@@ -4,7 +4,7 @@ from flask import jsonify, request
 from flask.typing import ResponseReturnValue
 from flask.views import MethodView
 
-from modules.access_token.rest_api.access_auth_middleware import access_auth_middleware
+from modules.authentication.rest_api.access_auth_middleware import access_auth_middleware
 from modules.account.account_service import AccountService
 from modules.account.types import (
     AccountSearchByIdParams,
