+++ src/apps/backend/modules/authentication/rest_api/access_auth_middleware.py
@@ -3,8 +3,8 @@ from typing import Any, Callable
 
 from flask import request
 
-from modules.access_token.access_token_service import AccessTokenService
-from modules.access_token.errors import (
+from modules.authentication.authentication_service import AuthenticationService
+from modules.authentication.errors import (
     AuthorizationHeaderNotFoundError,
     InvalidAuthorizationHeaderError,
     UnauthorizedAccessError,
@@ -22,7 +22,7 @@ def access_auth_middleware(next_func: Callable) -> Callable:
         if auth_scheme != "Bearer" or not auth_token:
             raise InvalidAuthorizationHeaderError("Invalid authorization header.")
 
-        auth_payload = AccessTokenService.verify_access_token(token=auth_token)
+        auth_payload = AuthenticationService.verify_access_token(token=auth_token)
 
         if "account_id" in kwargs and auth_payload.account_id != kwargs["account_id"]:
             raise UnauthorizedAccessError("Unauthorized access.")
