+++ src/apps/backend/modules/authentication/rest_api/authentication_router.py
@@ -1,9 +1,9 @@
 from flask import Blueprint
 
-from modules.access_token.rest_api.access_token_view import AccessTokenView
+from modules.authentication.rest_api.access_token_view import AccessTokenView
 
 
-class AccessTokenRouter:
+class AuthenticationRouter:
     @staticmethod
     def create_route(*, blueprint: Blueprint) -> Blueprint:
         blueprint.add_url_rule("/access-tokens", view_func=AccessTokenView.as_view("access_token_view"))
