+++ tests/modules/account/test_account_service.py
@@ -1,6 +1,6 @@
 from unittest.mock import patch
 
-from modules.access_token.types import AccessTokenPayload
+from modules.authentication.types import AccessTokenPayload
 from modules.account.account_service import AccountService
 from modules.account.errors import AccountNotFoundError
 from modules.account.types import (
@@ -26,7 +26,7 @@ class TestAccountService(BaseTestAccount):
         assert account.first_name == "first_name"
         assert account.last_name == "last_name"
 
-    @patch("modules.access_token.access_token_service.AccessTokenService.verify_access_token")
+    @patch("modules.authentication.authentication_service.AuthenticationService.verify_access_token")
     def test_get_account_by_id(self, mock_verify_access_token) -> None:
         account = AccountService.create_account_by_username_and_password(
             params=CreateAccountByUsernameAndPasswordParams(
@@ -43,7 +43,7 @@ class TestAccountService(BaseTestAccount):
         assert get_account_by_id.first_name == account.first_name
         assert get_account_by_id.last_name == account.last_name
 
-    @patch("modules.access_token.access_token_service.AccessTokenService.verify_access_token")
+    @patch("modules.authentication.authentication_service.AuthenticationService.verify_access_token")
     def test_throw_exception_when_usernot_exist(self, mock_verify_access_token) -> None:
         try:
             mock_verify_access_token.return_value = AccessTokenPayload(account_id="5f7b1b7b4f3b9b1b3f3b9b1b")
